chore: Consolidate search related code into the same package.
This makes the data layer package structure a bit cleaner and more organized.
