+++ data/src/main/kotlin/com/routesearch/data/search/SearchResults.kt
@@ -1,9 +1,18 @@
 package com.routesearch.data.search
 
-import com.routesearch.data.area.search.AreaSearchResult
-import com.routesearch.data.climb.search.ClimbSearchResult
-
 data class SearchResults(
   val areaSearchResults: List<AreaSearchResult>,
   val climbSearchResults: List<ClimbSearchResult>,
+)
+
+data class ClimbSearchResult(
+  val id: String,
+  val name: String,
+)
+
+data class AreaSearchResult(
+  val id: String,
+  val name: String,
+  val pathTokens: List<String>,
+  val totalClimbs: Int,
 )
