+++ features/src/main/kotlin/com/routesearch/features/search/SearchViewState.kt
@@ -1,7 +1,7 @@
 package com.routesearch.features.search
 
-import com.routesearch.data.area.search.AreaSearchResult
-import com.routesearch.data.climb.search.ClimbSearchResult
+import com.routesearch.data.search.AreaSearchResult
+import com.routesearch.data.search.ClimbSearchResult
 
 internal data class SearchViewState(
   val areaSearchResults: List<AreaSearchResult> = emptyList(),
